requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
celery==5.3.0
redis==4.5.4
pillow==9.5.0
requests==2.31.0
